+++ assets/scripts/items/events-items.js
@@ -23,21 +23,6 @@ const addNewItem = function (event) {
   }
 }
 
-const fillEditInputs = function (event) {
-  let itemId = $(event.target).closest('tbody').data('id')
-  store.id = itemId
-  let itemName = $(event.target).closest('tbody').data('name')
-  let itemDescription = $(event.target).closest('tbody').data('description')
-  let itemLocation = $(event.target).closest('tbody').data('location')
-  let itemPriority = $(event.target).closest('tbody').data('priority')
-  let itemStatus = $(event.target).closest('tbody').data('status')
-  console.log(itemName, itemDescription)
-  $('#editItem input[name="items[name]"]').val(itemName)
-  $('#editItem input[name="items[description]"]').val(itemDescription)
-  $('#editItem input[name="items[location]"]').val(itemLocation)
-  $('#editItem input[name="items[priority]"]').val(itemPriority)
-}
-
 // See all items
 const onShowItems = function () {
   api.showItems()
@@ -49,12 +34,13 @@ const onEditItem = function (event) {
   event.preventDefault()
   const data = getFormFields(event.target)
   if (data.items.priority < 1 || data.items.priority > 10) {
+    ui.resetInputFields()
   } else {
     api.updateItem(data)
       .then(ui.closeAllModals)
       .then(api.selectItem)
       .then(ui.selectItemSuccess)
-      .catch(console.log('fail'))
+      .catch(ui.editItemFail)
   }
 }
 
@@ -79,7 +65,7 @@ const onSelectItem = function () {
 const itemsEventHandler = function () {
   $('#item-form').on('submit', addNewItem)
   $('#bucket-list-content').on('click', 'button.delete-button', onDeleteItem)
-  $('#bucket-list-content').on('click', 'button.edit-button', fillEditInputs)
+  $('#bucket-list-content').on('click', 'button.edit-button', ui.fillEditInputs)
   $('#bucket-list-content').on('click', 'button.back-button', onShowItems)
   $('#bucket-list-content').on('click', 'td.item-name', onSelectItem)
   $('#edit-item-form').on('submit', onEditItem)
