+++ assets/scripts/items/ui-items.js
@@ -2,6 +2,7 @@
 
 const listTemplate = require('../../templates/list.handlebars')
 const listItemTemplate = require('../../templates/individual-item.handlebars')
+const store = require('../store.js')
 
 const newItemSuccess = function () {
 }
@@ -19,6 +20,11 @@ const showItemsSuccess = function (data) {
 }
 
 const showItemsFail = function () {
+}
+
+const editItemFail = function () {
+  console.log('fail')
+  resetInputFields()
 }
 
 const deleteItemSuccess = function () {
@@ -42,6 +48,32 @@ const closeAllModals = function () {
   $('.modal-backdrop').remove()
 }
 
+const fillEditInputs = function (event) {
+  let itemId = $(event.target).closest('tbody').data('id')
+  store.id = itemId
+  let itemName = $(event.target).closest('tbody').data('name')
+  store.name = itemName
+  let itemDescription = $(event.target).closest('tbody').data('description')
+  store.description = itemDescription
+  let itemLocation = $(event.target).closest('tbody').data('location')
+  store.location = itemLocation
+  let itemPriority = $(event.target).closest('tbody').data('priority')
+  store.priority = itemPriority
+  let itemStatus = $(event.target).closest('tbody').data('status')
+  store.status = itemStatus
+  $('#editItem input[name="items[name]"]').val(itemName)
+  $('#editItem input[name="items[description]"]').val(itemDescription)
+  $('#editItem input[name="items[location]"]').val(itemLocation)
+  $('#editItem input[name="items[priority]"]').val(itemPriority)
+}
+
+const resetInputFields = function () {
+  $('#editItem input[name="items[name]"]').val(store.name)
+  $('#editItem input[name="items[description]"]').val(store.description)
+  $('#editItem input[name="items[location]"]').val(store.location)
+  $('#editItem input[name="items[priority]"]').val(store.priority)
+}
+
 module.exports = {
   newItemSuccess,
   newItemFail,
@@ -50,5 +82,8 @@ module.exports = {
   deleteItemSuccess,
   deleteItemFail,
   selectItemSuccess,
-  closeAllModals
+  closeAllModals,
+  fillEditInputs,
+  editItemFail,
+  resetInputFields
 }
