+++ assets/styles/index.scss
@@ -43,7 +43,7 @@ padding: 2rem;
   display: none;
 }
 
-.to-signup {
+#to-signup {
   text-align: center;
 }
